Fix for missed prices on product-variation buttons.
This is a crude (and maybe buggy) fix for the prices on non-selected
product-variation buttons. I'm both learning a new language/environment
and getting rather tired, so I'm definitely not such if it's the right
approach. Though notably, these newly handled prices are keeping their
format preserved (the span is staying intact), so it makes me wonder if
the overall algorithm needs to be generalized so as to only replace the
intended text.

Example:

http://www.amazon.com/Cuisinart-CSB-75BC-2-Speed-Immersion-Blender/dp/B00ARQVM5O/